Substitute variables on export
* ob-http.el (ob-http-export-expand-variables): New function handling
  variable substitution.

To use this functionality, the function has to be added to Org export
hook (org-export-before-processing-hook) before the source blocks have
been executed.

It might be possible to hook it up somewhere in `ob-http' internals, but
this seems like a cleaner solution. And it should not break anything by
itself (until it gets added to the hook, that is).

References issue #25.